requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
numpy
